Return null from services in case of errors. Handle null in controllers

+++ src/main/java/com/KoAB91/geology/service/GeoClassService.java
@@ -2,6 +2,7 @@ package com.KoAB91.geology.service;
 
 import com.KoAB91.geology.dto.GeoClassDTO;
 import com.KoAB91.geology.entity.GeologicalClass;
+import com.KoAB91.geology.enums.DeleteResponse;
 import com.KoAB91.geology.extensions.StreamExtension;
 import com.KoAB91.geology.repository.GeologicalClassRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +20,9 @@ public class GeoClassService {
     public List<GeoClassDTO> get() {
         List<GeologicalClass> geoClasses = StreamExtension.toList(
                 geologicalClassRepository.findAll());
-
+        if (geoClasses == null){
+            return null;
+        }
         List<GeoClassDTO> geoClassesDTO = geoClasses
                 .stream()
                 .map(this::convertToDTO)
@@ -30,17 +33,31 @@ public class GeoClassService {
 
     public GeoClassDTO getByName(String name) {
         GeologicalClass geoClass = geologicalClassRepository.findByName(name);
+        if (geoClass == null) {
+            return null;
+        }
         return convertToDTO(geoClass);
 
     }
 
     public GeoClassDTO getByCode(String name) {
         GeologicalClass geoClass = geologicalClassRepository.findByCode(name);
+        if (geoClass == null) {
+            return null;
+        }
         return convertToDTO(geoClass);
     }
 
     public GeoClassDTO create(String className, String classCode) {
-        GeologicalClass geoClass = new GeologicalClass();
+        GeologicalClass geoClass = geologicalClassRepository.findByName(className);
+        if (geoClass != null){
+            return convertToDTO(geoClass);
+        }
+        geoClass = geologicalClassRepository.findByCode(classCode);
+        if (geoClass != null){
+            return convertToDTO(geoClass);
+        }
+        geoClass = new GeologicalClass();
         geoClass.setName(className);
         geoClass.setCode(classCode);
         geoClass = geologicalClassRepository.save(geoClass);
@@ -50,25 +67,34 @@ public class GeoClassService {
     public GeoClassDTO updateName(String name, String newName) {
 
         GeologicalClass geoClass = geologicalClassRepository.findByName(name);
+        if (geoClass == null){
+            return null;
+        }
         geoClass.setName(newName);
         geologicalClassRepository.save(geoClass);
 
         return convertToDTO(geoClass);
     }
 
-    public GeoClassDTO updateCode(String name, String newCode) {
+    public GeoClassDTO updateCode(String code, String newCode) {
 
-        GeologicalClass geoClass = geologicalClassRepository.findByName(name);
+        GeologicalClass geoClass = geologicalClassRepository.findByCode(code);
+        if (geoClass == null){
+            return null;
+        }
         geoClass.setCode(newCode);
         geologicalClassRepository.save(geoClass);
 
         return convertToDTO(geoClass);
     }
 
-    public void delete(String name) {
+    public DeleteResponse delete(String name) {
         GeologicalClass geoClass = geologicalClassRepository.findByName(name);
-        // geoClass.deleteGeoCLassFromSections();
+        if (geoClass == null){
+            return DeleteResponse.NOT_FOUND;
+        }
         geologicalClassRepository.delete(geoClass);
+        return DeleteResponse.OK;
     }
 
     private GeoClassDTO convertToDTO(GeologicalClass geoClass) {
