+++ src/main/java/com/KoAB91/geology/service/SectionService.java
@@ -4,6 +4,7 @@ import com.KoAB91.geology.dto.GeoClassDTO;
 import com.KoAB91.geology.dto.SectionDTO;
 import com.KoAB91.geology.entity.GeologicalClass;
 import com.KoAB91.geology.entity.Section;
+import com.KoAB91.geology.enums.DeleteResponse;
 import com.KoAB91.geology.extensions.StreamExtension;
 import com.KoAB91.geology.repository.GeologicalClassRepository;
 import com.KoAB91.geology.repository.SectionRepository;
@@ -25,6 +26,9 @@ public class SectionService {
     public List<SectionDTO> get() {
         List<Section> sections = StreamExtension.toList(
                 sectionRepository.findAll());
+        if (sections == null) {
+            return null;
+        }
 
         List<SectionDTO> sectionsDTO = sections
                 .stream()
@@ -36,11 +40,18 @@ public class SectionService {
 
     public SectionDTO getByName(String name) {
         Section section = sectionRepository.findByName(name);
+        if (section == null) {
+            return null;
+        }
         return convertToDTO(section);
     }
 
     public List<SectionDTO> getByСlassCode(String classCode) {
         GeologicalClass geoClass = geologicalClassRepository.findByCode(classCode);
+        if (geoClass == null) {
+            return null;
+        }
+
         List<SectionDTO> sectionsDTO = geoClass.getSections()
                 .stream()
                 .map(this::convertToDTO)
@@ -51,14 +62,20 @@ public class SectionService {
 
 
     public SectionDTO create(String name) {
-        Section section = new Section();
-        section.setName(name);
-        section = sectionRepository.save(section);
+        Section section = sectionRepository.findByName(name);
+        if (section == null) {
+            section = new Section();
+            section.setName(name);
+            section = sectionRepository.save(section);
+        }
         return convertToDTO(section);
     }
 
     public SectionDTO updateName(String name, String newName) {
         Section section = sectionRepository.findByName(name);
+        if (section == null) {
+            return null;
+        }
         section.setName(newName);
         section = sectionRepository.save(section);
         return convertToDTO(section);
@@ -66,15 +83,26 @@ public class SectionService {
 
     public SectionDTO addGeoClass(String name, String className) {
         Section section = sectionRepository.findByName(name);
+        if (section == null) {
+            return null;
+        }
         GeologicalClass geoClass = geologicalClassRepository.findByName(className);
+        if (geoClass == null) {
+            return null;
+        }
         section.getGeologicalClasses().add(geoClass);
         section = sectionRepository.save(section);
         return convertToDTO(section);
     }
 
-    public void delete(String name) {
+    public DeleteResponse delete(String name) {
         Section section = sectionRepository.findByName(name);
+        if (section == null) {
+            return DeleteResponse.NOT_FOUND;
+        }
         sectionRepository.delete(section);
+        return DeleteResponse.OK;
+
     }
 
     private SectionDTO convertToDTO(Section section) {
